Updated tito.properties with new upload endpoint.
Renamed the zoidberg.getuserhomedir and zoidberg.fileupload services properties to include their service names instead of zoidberg.
Updated the sample host properties to include the service subdir, instead of hard-coding those in the services properties, since some services may simply reside on a port and not under a URL subdir.