Fixed NewToolRequestServlet::executeAction bugs.
The jsonInfo object is intended only for the email message that is sent, and any errors that occur in UploadServlet will be populated in jsonErrors object, so those results are now returned by NewToolRequestServlet::executeAction.
Removed reference to org.iplantc.js.integrate.client.I18N, since GWT can only create this instance in the client and will throw an exception in this servlet.