CORE-3012 Merged Save and Publish buttons.
Removed the old Save button and replaced it with the Publish button, which is now labeled as "Save" and uses the save icon.
It still saves and publishes the App, but it also updates both the last edited and published dates.
Now a successful publish will fire an AnalysisSelectEvent for that App instead of displaying a popup message with a link.
Removed some obsolete display strings.