CORE-3079 Added orderingRequired util method.
Added a PropertyUtil class with a static orderingRequired method that checks if the given property should be ordered in the CommandLineOrderingGridPanel.
Updated the CommandLineOrderingGridPanel and the TemplateTabPanel publish action to verify property ordering with this util method.
Refactored the unordered-publish-confirmation dialog in TemplateTabPanel into a confirmUnorderedPublish method.
Refactored the TYPE_STATIC_TEXT constant from NavigationTreePanel into the new PropertyUtil class.