CORE-3533 Fixed to re-enable EditableNumberList.
Since the AbstractListPropertyEditorPanel::allocateList method is called in the constructor for PropertyTypeEditorPanel, it was being called before the category was set in the constructor for PropertyValidationEditorPanel.
Fixed this bug by creating Number and String ListboxEditorPanels that build the correct BasicEditableList for their allocateList methods.
Refactored Number (de)serializing to/from JSON in EditableNumberList to avoid NPEs.