CORE-3517 Integrated TreeSelection Property support.
Added support for TreeSelection types to PropertyValidationEditorPanel and StringPropertyEditorPanel.
Added a TreeListEditorPanel that wraps the ListRuleArgumentEditor GXT3 widget in a custom WidgetComponent that implements the ListEditor interface.
Updated ListRuleArgumentEditor to be able to get and set a ListRuleArgumentGroup of the values in its grid.
Added a ListRuleArgumentEditor::addStoreHandlers method to allow the TreeListEditorPanel to set event handlers to update its property.