CORE-3517 Added Check Cascade config to ListRuleArgumentEditor.
Added a combo box to the ListRuleArgumentEditor toolbar to allow the Check Cascade to be set for the tree selector.
This allows only a group to be selected (and not all of its children), so groups may now have arguments and values.
Refactored the ListRuleArgumentEditor toolbar so that the buttons have only tooltips and not text, to make room for the combo box.
Added a folder_add.png icon for the Add Group action.